ui: mobile/responsive hardening, bundle perf, and UX polish
Responsive/mobile:
- sheet/dialog full-width + max-h on phones; TaskDetailPanel + sidebar width clamps; index.html viewport-fit=cover + safe-area-inset padding; 44px touch targets on touched controls

Performance:
- replace 'import * as PhosphorIcons' with a curated icon map (icon chunk ~5MB -> ~421KB); lazy-load heavy _app routes via extracted screen components + RouteFallback (main chunk ~1.84MB -> ~1.46MB)

UX:
- onboarding: labelled progress for a11y, surface handleFinish errors with retry, dependency-free password-strength meter
- replace window.confirm in ChatScreen + SchedulesList with a new design-system AlertDialog (built on existing Radix Dialog, no new dep)
