+++ index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0, viewport-fit=cover" />
     <title>Omnipus — Elite Simplicity. Sovereign Control.</title>
     <meta name="description" content="Omnipus is the autonomous command center that provides elite feature-richness with military-grade security." />
