+++ src/components/chat/historical-markdown.tsx
@@ -16,12 +16,15 @@ import { rewriteLegacyURL, resolveEffectivePreview } from '@/lib/preview-url'
 import { isSafeHref } from '@/lib/url-safe'
 import { fetchAboutInfo } from '@/lib/api'
 import { ImageLightbox } from './image-lightbox'
-import * as PhosphorIcons from '@phosphor-icons/react'
+import { PHOSPHOR_EMOJI_ICONS } from '@/lib/phosphor-emoji-icons'
 import type { ComponentPropsWithoutRef } from 'react'
 
+// Icons resolved from an explicit allow-list (PHOSPHOR_EMOJI_ICONS) — never a
+// wildcard `import * as` (which pulls the whole ~5MB icon set). See
+// src/lib/phosphor-emoji-icons.tsx. Mirrors markdown-text.tsx.
 function PhosphorEmojiSpan({ 'data-phosphor-icon': iconName, children, ...props }: ComponentPropsWithoutRef<'span'> & { 'data-phosphor-icon'?: string }) {
-  if (iconName && iconName in PhosphorIcons) {
-    const Icon = (PhosphorIcons as unknown as Record<string, React.ComponentType<{ size?: number; weight?: string; className?: string }>>)[iconName]
+  const Icon = iconName ? PHOSPHOR_EMOJI_ICONS[iconName] : undefined
+  if (Icon) {
     return <Icon size={14} weight="regular" className="inline-block align-middle text-[var(--color-accent)] mx-0.5" />
   }
   return <span {...props}>{children}</span>
