+++ src/styles/globals.css
@@ -42,7 +42,8 @@
   --font-mono: 'JetBrains Mono', ui-monospace, monospace;
 
   /* Spacing scale additions */
-  --spacing-sidebar: 256px;
+  /* Clamp so a pinned/overlaid sidebar never exceeds the viewport on small screens */
+  --spacing-sidebar: min(256px, 80vw);
 
   /* Animation */
   --ease-spring: cubic-bezier(0.34, 1.56, 0.64, 1);
@@ -94,4 +95,31 @@ body {
 }
 ::-webkit-scrollbar-thumb:hover {
   background: var(--color-muted);
+}
+
+/* ------------------------------------------------------------------ */
+/* Safe-area handling (viewport-fit=cover)                             */
+/* With viewport-fit=cover the app shell extends edge-to-edge under    */
+/* notches and home indicators. Pad the shell chrome with             */
+/* env(safe-area-inset-*) so content is never hidden. The sidebar      */
+/* applies its own insets inline (see Sidebar.tsx). The app-shell      */
+/* root is the only top-level dvh flex container; its <header> top bar */
+/* and <main> content column are targeted structurally so no JSX       */
+/* change is required.                                                 */
+/* ------------------------------------------------------------------ */
+
+/* App-shell root reserves the right notch inset for the whole shell. */
+.h-dvh {
+  padding-right: env(safe-area-inset-right);
+}
+
+/* Top bar — clear the top notch (max() keeps the existing py-3 when
+   there is no notch). */
+.h-dvh header {
+  padding-top: max(0.75rem, env(safe-area-inset-top));
+}
+
+/* Bottom home-indicator inset on the scrollable content column. */
+.h-dvh main {
+  padding-bottom: env(safe-area-inset-bottom);
 }
